+++ src/documentation/webview/CommunicationBridge.ts
@@ -16,7 +16,7 @@ import { Disposable } from "./Disposable";
 /**
  * Sends and receives messages from swift-docc-render
  */
-export interface CommunicationBridge {
+interface CommunicationBridge {
     send(message: VueAppMessage): void;
     onDidReceiveMessage(handler: (message: VueAppMessage) => void): Disposable;
 }
@@ -83,13 +83,13 @@ export function createCommunicationBridge(): Promise<CommunicationBridge> {
 /**
  * Represents a message that can be sent between the webview and swift-docc-render
  */
-export type VueAppMessage = RenderedMessage | NavigationMessage | UpdateContentMessage;
+type VueAppMessage = RenderedMessage | NavigationMessage | UpdateContentMessage;
 
 /**
  * Sent from swift-docc-render to the webview when content as been rendered
  * to the screen.
  */
-export interface RenderedMessage {
+interface RenderedMessage {
     type: "rendered";
 }
 
@@ -100,7 +100,7 @@ export interface RenderedMessage {
  * need to send an {@link UpdateContentMessage} after the first render to
  * switch pages.
  */
-export interface NavigationMessage {
+interface NavigationMessage {
     type: "navigation";
     data: string;
 }
@@ -113,7 +113,7 @@ export interface NavigationMessage {
  * JavaScript object before sending. Raw strings will not be parsed
  * automatically.
  */
-export interface UpdateContentMessage {
+interface UpdateContentMessage {
     type: "contentUpdate";
     data: unknown;
 }