+++ src/utilities/commands.ts
@@ -16,6 +16,5 @@ export enum Workbench {
     ACTION_DEBUG_CONTINUE = "workbench.action.debug.continue",
     ACTION_CLOSEALLEDITORS = "workbench.action.closeAllEditors",
     ACTION_RELOADWINDOW = "workbench.action.reloadWindow",
-    ACTION_PREVIOUSEDITORINGROUP = "workbench.action.previousEditorInGroup",
     ACTION_QUIT = "workbench.action.quit",
 }