+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 
 - Swift-testing test runs are marked as 'started' in the UI immediately, not after compilation finishes ([#2079](https://github.com/swiftlang/vscode-swift/pull/2079))
+- Fix legacy boolean setting values for `swift.sourcekit-lsp.backgroundIndexing` not being recognized ([#2092](https://github.com/swiftlang/vscode-swift/pull/2092))
 
 ## 2.16.1 - 2026-02-02